plugins

+++ gwa-developer-key/src/main/angular/src/app/Service/BaseService.ts
@@ -1,3 +1,4 @@
+import { Observable } from 'rxjs/Observable';
 import 'rxjs/add/operator/toPromise';
 import {
   Injectable,
@@ -11,6 +12,7 @@ import {
 import {
   Headers, 
   Http,
+  Response,
   URLSearchParams
 } from '@angular/http';
 
@@ -41,9 +43,9 @@ export abstract class BaseService<T> implements Service<T> {
 
   dialog: MdDialog = this.injector.get(MdDialog);
 
-  private jsonHeaders = {
-    headers: new Headers({ 'Content-Type': 'application/json' })
-  };
+  usePostForDelete : boolean = true;
+  
+  private jsonHeaders =  new Headers({ 'Content-Type': 'application/json' });
 
   constructor(
     protected injector : Injector,
@@ -66,7 +68,7 @@ export abstract class BaseService<T> implements Service<T> {
     return this.http.post(
       url,
       jsonText,
-      this.jsonHeaders
+      { headers: this.jsonHeaders }
     ).toPromise()
       .then(response => {
         const json = response.json();
@@ -125,13 +127,29 @@ export abstract class BaseService<T> implements Service<T> {
     }
 
     const url = this.getUrl(path);
-    return this.http.delete(
-      url,
-      {
-        headers: new Headers({ 'Content-Type': 'application/json' }),
-        search: params
-      }
-    ).toPromise()
+    var response : Observable<Response>;
+    if (this.usePostForDelete) {
+      response = this.http.post(
+        url,
+        '',
+        {
+          headers: new Headers({
+            'Content-Type': 'application/json',
+            'X-HTTP-Method-Override': 'DELETE'
+          }),
+          search: params
+        }
+      );
+    } else {
+      response = this.http.delete(
+        url,
+        {
+          headers: this.jsonHeaders,
+          search: params
+        }
+      );
+    }
+    return response.toPromise()
       .then(response => {
         const json = response.json();
         if (json.error) {
@@ -222,16 +240,18 @@ export abstract class BaseService<T> implements Service<T> {
   getRowsPage(
     offset: number,
     limit: number,
-    filterFieldName : string,
-    filterValue : string,
-    path: string
+    path: string,
+    filter : { [fieldName: string] : string}
   ): Promise<any> {
     let params = new URLSearchParams();
     params.set('offset', offset.toString()); 
     params.set('limit', limit.toString());
-    if (filterFieldName && filterValue) {
-      params.set("filterFieldName", filterFieldName);
-      params.set("filterValue", filterValue);
+    if (filter) {
+      for (const fieldName in filter) {
+        const value = filter[fieldName];
+        params.set("filterFieldName", fieldName);
+        params.set("filterValue", value);
+      }
     }
     if (!path) {
       path = this.path;
@@ -292,7 +312,7 @@ export abstract class BaseService<T> implements Service<T> {
     return this.http.put(
       url,
       jsonText,
-      this.jsonHeaders
+      { headers: this.jsonHeaders }
     ).toPromise()
       .then(response => {
         const json = response.json();
